+++ ts/types/Decoder.d.ts
@@ -5,25 +5,6 @@ import { Codec } from "./Codec"
 import { CodecContext, CodecContext_base } from "./CodecContext"
 import { Demuxer } from "./Demuxer"
 
-/** The DecodedFrames object is returned as the result of a decode operation */
-export interface DecodedFrames {
-	/** Object name. */
-	readonly type: 'frames'
-	/** 
-	   * Decoded frames that are now available. If the array is empty, the decoder has buffered
-	 * the packet as part of the process of producing future frames
-	   */
-	readonly frames: Array<Frame>
-	/** Total time in microseconds that the decode operation took to complete */
-	readonly total_time: number
-
-
-	timings?: { 
-		// "encode" | "dice" | "decode" | "filter"
-		[key: string]: Timing; };
-
-}
-
 export interface Decoder extends CodecContext_base {
 	// readonly type: 'decoder'
 	// readonly time_base: [number, number]
